requirements.txt
fastapi
uvicorn
sqlalchemy
tensorflow-cpu
opencv-python-headless
numpy
python-multipart